+++ tests/trio_options.py
@@ -1,0 +1,3 @@
+@app.route  # type: ignore
+async def f():
+    print("hello world")