+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 *[CalVer, YY.month.patch](https://calver.org/)*
 
+## Future
+- Added `--no-checkpoint-warning-decorators`.
+
 ## 22.8.8
 - Fix false alarm on TRIO107 with checkpointing `try` and empty `finally`
 - Fix false alarm on TRIO107&108 with infinite loops